Avoid acquiring lock when already locked (#3)
The setex Redis command that we're using currently will, as noted, Set the time to live in seconds of a key

It does not guarantee uniqueness of the key (in fact, if you call setex again on the same key, it will only update the TTL)

The proposed fix is to use the set command, which by description Set the string value of a key, but when used with the options :ex and :nx give us the desired behaviour:

  #   - `:ex => Fixnum`: Set the specified expire time, in seconds.
  #   - `:nx => true`: Only set the key if it does not already exist.

Additionally, we will also check the return value of the set command. When it returns false it means a lock could not be acquired (already locked), and we can then raise an error in the lock! method