Compact arguments for default lock_key
* Objective

This commit objective is about compacting the arguments while generating
the lock_key.

* Why

If nil agruments are added to the job execution it does change the MD5,
and it can cause a likely similar job to have a different lock_key and
bypass the locking.

* How

This commit compacts the arguments while generating the default
lock_key.